fixed a bug where numpy array dtype wouldnt be set correctly

+++ esiaf_wav_player/wav_player.py
@@ -46,7 +46,7 @@ class WavPlayer():
         frames_per_second = 16000
         timeout_in_seconds = 2. * 400./1000.
         five_twelve_blocks = int(numpy.ceil(frames_per_second * timeout_in_seconds/512))
-        filler_array = numpy.zeros((512,), numpy.int16)
+        filler_array = numpy.zeros((512,), numpy.float64)
 
         for _ in range(five_twelve_blocks):
             end_time = self.start_time + self.frametime
